Uses minio charm instead of microk8s add-on Adds helpers file for itests Adds a minor fix in unit tests
