+++ tox.ini
@@ -51,11 +51,6 @@ commands =
     coverage report
 
 [testenv:integration]
-passenv =
-    {[testenv]pass_env}
-    S3_ACCESS_KEY
-    S3_SECRET_KEY
-    S3_ENDPOINT
 description = Run integration tests
 commands =
     pytest --asyncio-mode=auto -v --tb native {[vars]integration_test_path} --log-cli-level=INFO -s {posargs}
